made definition_parser into a class, and added a unit test
